Reorganize behavior specs
the `behavior.spec.js` file itself needs additional clean up to match the API, but it is now the correct file and has coverage for `behavior.js`
